Notification is now its own component
