+++ src/components/AccountForm/AccountCreate/AccountCreate.js
@@ -1,16 +1,15 @@
-import React, { useState, useRef } from 'react';
+import React, {  useRef } from 'react';
 import { Link, useHistory } from 'react-router-dom';
 import '../AccountForm.css';
 
 import { useDispatch } from 'react-redux';
 import { accountCreate, accountLogin } from '../../../reducers/UserSlice';
+import { setErrorMessage } from '../../../reducers/NotificationSlice';
 
 
 function AccountCreate() {
   const history = useHistory();
   const dispatch = useDispatch();
-
-  const [errorMessage, setErrorMessage] = useState(null);
 
   const email = useRef();
   const firstName = useRef();
@@ -24,7 +23,7 @@ function AccountCreate() {
    */
   async function accountCreateSubmit() {
     if (password.current.value !== passwordVerify.current.value) {
-      setErrorMessage('Passwords do not match!');
+      dispatch(setErrorMessage('Passwords do not match!'));
       return;
     }
     
@@ -60,7 +59,7 @@ function AccountCreate() {
       if (err && 'message' in err) {
         errorMessage = err['message'];
       }
-      setErrorMessage(errorMessage);
+      dispatch(setErrorMessage(errorMessage));
     }
   }
 
@@ -109,17 +108,6 @@ function AccountCreate() {
     );
   }
 
-  function errorMessageDisplay() {
-    if (errorMessage) {
-      return (
-        <div className="account-form-error-area">
-          <p className="account-form-error-msg">{errorMessage}</p>
-        </div>
-      );
-    }
-  }
-
-
   return (
     <div className="account-form-container" onKeyPress={keyboardFormSubmit}>
       <div className="account-form-header">
@@ -127,7 +115,6 @@ function AccountCreate() {
       </div>
       {createForm()}
       {createActions()}
-      {errorMessageDisplay()}
     </div>
   );
 }
