+++ src/components/Notification/index.js
@@ -1,0 +1,2 @@
+import Notification from './Notification';
+export default Notification;
