+++ src/containers/App.js
@@ -7,42 +7,14 @@ import {
 import './App.css';
 import './Landing.css';
 
-import { Snackbar, IconButton }from '@material-ui/core';
-import { Close } from '@material-ui/icons';
-
-import { useDispatch, useSelector } from 'react-redux';
-import { clearErrorMessage } from '../reducers/NotificationSlice';
-
 import Home from '../components/Home';
 import AccountForm from '../components/AccountForm';
 import RoomForm from '../components/RoomForm';
 import Room from '../components/Room';
+import Notification from '../components/Notification';
 
 
 function App() {
-  const dispatch = useDispatch();
-  const errorMessage = useSelector(state => (state.notification.error));
-
-  function displayNotification() {
-    return (
-      <Snackbar
-        anchorOrigin={{ vertical: 'bottom', horizontal: 'left' }}
-        open={errorMessage !== null}
-        autoHideDuration={5000}
-        onClose={() => {dispatch(clearErrorMessage())}}
-        message={errorMessage}
-        action={
-          <React.Fragment>
-            <IconButton size="small" aria-label="close" color="inherit" onClick={() => {dispatch(clearErrorMessage())}}>
-              <Close fontSize="small" />
-            </IconButton>
-          </React.Fragment>
-        }
-      />
-    );
-  }
-
-
   return (
     <div className="App">
       <div className="landing-container">
@@ -72,7 +44,7 @@ function App() {
             </Switch>
           </Router>
         </div>
-        {displayNotification()}
+        <Notification />
       </div>
     </div>
   );
